sync process
