+++ assertions/BUILD.bazel
@@ -3,8 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "assertions",
     srcs = [
+        "confirmation.go",
         "poster.go",
         "scanner.go",
+        "sync.go",
     ],
     importpath = "github.com/OffchainLabs/bold/assertions",
     visibility = ["//visibility:public"],
@@ -24,7 +26,6 @@ go_library(
         "@com_github_ethereum_go_ethereum//accounts/abi/bind",
         "@com_github_ethereum_go_ethereum//common",
         "@com_github_ethereum_go_ethereum//core/types",
-        "@com_github_ethereum_go_ethereum//crypto",
         "@com_github_ethereum_go_ethereum//log",
         "@com_github_ethereum_go_ethereum//metrics",
         "@com_github_pkg_errors//:errors",
@@ -39,25 +40,4 @@ go_test(
         "scanner_test.go",
     ],
     embed = [":assertions"],
-    deps = [
-        "//chain-abstraction:protocol",
-        "//challenge-manager",
-        "//challenge-manager/types",
-        "//containers/option",
-        "//containers/threadsafe",
-        "//layer2-state-provider",
-        "//solgen/go/mocksgen",
-        "//solgen/go/rollupgen",
-        "//testing",
-        "//testing/mocks",
-        "//testing/mocks/state-provider",
-        "//testing/setup:setup_lib",
-        "//util",
-        "@com_github_ethereum_go_ethereum//accounts/abi/bind",
-        "@com_github_ethereum_go_ethereum//common",
-        "@com_github_pkg_errors//:errors",
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//mock",
-        "@com_github_stretchr_testify//require",
-    ],
 )
