+++ assertions/poster.go
@@ -54,25 +54,26 @@ func (m *Manager) postAssertionRoutine(ctx context.Context) {
 }
 
 // PostAssertion differs depending on whether or not the validator is currently staked.
-func (m *Manager) PostAssertion(ctx context.Context) (option.Option[protocol.Assertion], error) {
+func (m *Manager) PostAssertion(ctx context.Context) (option.Option[*protocol.AssertionCreatedInfo], error) {
 	// Ensure that we only build on a valid parent from this validator's perspective.
 	// the validator should also have ready access to historical commitments to make sure it can select
 	// the valid parent based on its commitment state root.
 	m.assertionChainData.Lock()
 	parentAssertionCreationInfo, ok := m.assertionChainData.canonicalAssertions[m.assertionChainData.latestAgreedAssertion]
 	m.assertionChainData.Unlock()
+	none := option.None[*protocol.AssertionCreatedInfo]()
 	if !ok {
-		return option.None[protocol.Assertion](), fmt.Errorf(
+		return none, fmt.Errorf(
 			"latest agreed assertion %#x not part of canonical mapping, something is wrong",
 			m.assertionChainData.latestAgreedAssertion.Hash,
 		)
 	}
 	staked, err := m.chain.IsStaked(ctx)
 	if err != nil {
-		return option.None[protocol.Assertion](), err
+		return none, err
 	}
 	// If the validator is already staked, we post an assertion and move existing stake to it.
-	var assertionOpt option.Option[protocol.Assertion]
+	var assertionOpt option.Option[*protocol.AssertionCreatedInfo]
 	var postErr error
 	if staked {
 		assertionOpt, postErr = m.PostAssertionBasedOnParent(
@@ -85,10 +86,10 @@ func (m *Manager) PostAssertion(ctx context.Context) (option.Option[protocol.Ass
 		)
 	}
 	if postErr != nil {
-		return option.None[protocol.Assertion](), postErr
+		return none, postErr
 	}
 	if assertionOpt.IsSome() {
-		m.submittedAssertions.Insert(assertionOpt.Unwrap().Id().Hash)
+		m.submittedAssertions.Insert(assertionOpt.Unwrap().AssertionHash)
 	}
 	return assertionOpt, nil
 }
@@ -102,9 +103,10 @@ func (m *Manager) PostAssertionBasedOnParent(
 		parentCreationInfo *protocol.AssertionCreatedInfo,
 		newState *protocol.ExecutionState,
 	) (protocol.Assertion, error),
-) (option.Option[protocol.Assertion], error) {
+) (option.Option[*protocol.AssertionCreatedInfo], error) {
+	none := option.None[*protocol.AssertionCreatedInfo]()
 	if !parentCreationInfo.InboxMaxCount.IsUint64() {
-		return option.None[protocol.Assertion](), errors.New("inbox max count not a uint64")
+		return none, errors.New("inbox max count not a uint64")
 	}
 	// The parent assertion tells us what the next posted assertion's batch should be.
 	// We read this value and use it to compute the required execution state we must post.
@@ -118,9 +120,9 @@ func (m *Manager) PostAssertionBasedOnParent(
 					"batchCount": batchCount,
 				},
 			)
-			return option.None[protocol.Assertion](), nil
+			return none, nil
 		}
-		return option.None[protocol.Assertion](), errors.Wrapf(err, "could not get execution state at batch count %d", batchCount)
+		return none, errors.Wrapf(err, "could not get execution state at batch count %d", batchCount)
 	}
 	srvlog.Info(
 		"Posting assertion with retrieved state", log.Ctx{
@@ -133,11 +135,8 @@ func (m *Manager) PostAssertionBasedOnParent(
 		parentCreationInfo,
 		newState,
 	)
-	switch {
-	case errors.Is(err, solimpl.ErrAlreadyExists):
-		return option.None[protocol.Assertion](), errors.Wrap(err, "assertion already exists, was unable to post")
-	case err != nil:
-		return option.None[protocol.Assertion](), err
+	if err != nil {
+		return none, err
 	}
 	srvlog.Info("Submitted latest L2 state claim as an assertion to L1", log.Ctx{
 		"validatorName":         m.validatorName,
@@ -146,6 +145,9 @@ func (m *Manager) PostAssertionBasedOnParent(
 		"postedExecutionState":  fmt.Sprintf("%+v", newState),
 	})
 	assertionPostedCounter.Inc(1)
-
-	return option.Some(assertion), nil
+	creationInfo, err := m.chain.ReadAssertionCreationInfo(ctx, assertion.Id())
+	if err != nil {
+		return none, err
+	}
+	return option.Some(creationInfo), nil
 }
